Refactor Router and AI Assistant components for improved functionality
- Replaced CoursePlayerPage with CourseDetailPage and LessonPlayerPage in Router for better route management.
- Updated ProtectedRoute to public routes for lesson and course detail access.
- Enhanced AIAssistant, ChatPanel, and ChatSidebar components with a new disabled prop to control user interactions.
- Adjusted LessonList to include isLocked prop for lesson access control and improved lesson link handling.
- Removed CoursePlayerPage as part of the refactor.
