+++ src/features/ai-assistant/components/ChatSidebar.tsx
@@ -12,26 +12,30 @@ interface ChatSidebarProps {
   activeTopic: string | null;
   onSelectTopic: (topicId: string) => void;
   onNewChat: () => void;
+  disabled?: boolean;
 }
 
 export default function ChatSidebar({ 
   topics, 
   activeTopic, 
   onSelectTopic, 
-  onNewChat 
+  onNewChat, 
+  disabled = false
 }: ChatSidebarProps) {
   return (
     <div className="flex flex-col h-full bg-gray-50 border-r border-gray-200">
-      <div className="p-4 border-b border-gray-200 sticky top-0 bg-gray-50 z-10">
-        <Button 
-          onClick={onNewChat}
-          className="w-full flex items-center justify-center gap-2 bg-primary hover:bg-primary/90"
-          variant="default"
-        >
-          <Plus className="h-4 w-4" />
-          <span>New Chat</span>
-        </Button>
-      </div>
+      {!disabled && (
+        <div className="p-4 border-b border-gray-200 sticky top-0 bg-gray-50 z-10">
+          <Button 
+            onClick={onNewChat}
+            className="w-full flex items-center justify-center gap-2 bg-primary hover:bg-primary/90"
+            variant="default"
+          >
+            <Plus className="h-4 w-4" />
+            <span>New Chat</span>
+          </Button>
+        </div>
+      )}
       
       <div className="overflow-y-auto flex-1 py-2">
         {topics.length === 0 ? (
